+++ lib/config/createUrc.test.js
@@ -3,7 +3,7 @@
 const path = require('path');
 
 const getDefaultConfig = require('../default-underreact.config');
-const createConfig = require('./');
+const createConfig = require('./createUrc');
 
 test('normalizes empty configuration', () => {
   const defaultConfig = getDefaultConfig({
